update to use run parameters, some path checks
Valve did recently changes to the steam://connect/ command of their Steam Browser Protocol in a way that renders it non-functional. Perhaps the syntax changed and they forgot to update the documentation, idk, it just doesn't work anymore. So we are not using it anymore and instead use the Steam Client parameters to start the game with them.